Initial Commit
Definitely should have done this earlier, but was focused on getting it up and running first. The form connects to a Gateway API which can added the data into a DynamoDB database I have set up and email the message creator using SES (however it currently only works for other emails which SES has verified, since my account is still in sandbox mode).

I need to update the look of the form, since it is pretty much the bare minimum at the moment, and will also start looking into the doing the bonus objectives since I have some time left.